requirements.txt
alembic==0.7.7
appdirs==1.4.3
bcolz==0.12.1
bleach==1.5.0
Bottleneck==1.2.1
click==6.7
colorama==0.3.9
contextlib2==0.5.5
cycler==0.10.0
cyordereddict==0.2.2
Cython==0.26
decorator==4.1.2
empyrical==0.2.2
entrypoints==0.2.3
ez-setup==0.9
html5lib==0.9999999
intervaltree==2.1.0
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-console==5.1.0
jupyter-core==4.3.0
Logbook==0.12.5
lru-dict==1.1.4
Mako==1.0.6
MarkupSafe==0.23
matplotlib==2.0.2
mistune==0.7.4
multipledispatch==0.4.9
nbconvert==5.2.1
nbformat==4.3.0
networkx==1.11
notebook==5.0.0
numexpr==2.6.2
numpy==1.11.3
olefile==0.44
pandas==0.18.1
pandas-datareader==0.4.0
pandocfilters==1.4.1
path.py==10.3.1
patsy==0.4.1
pickleshare==0.7.4
Pillow==4.2.1
profilehooks==1.9.0
prompt-toolkit==1.0.14
protobuf==3.3.0
py==1.4.33
pyfolio==0.7.0
Pygments==2.2.0
pyparsing==2.2.0
pytest==3.1.1
python-dateutil==2.6.1
pytools==2017.1.1
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.3.0
requests==2.14.2
requests-file==1.4.1
requests-ftp==0.3.1
scikit-learn==0.18.1
scipy==0.19.0
seaborn==0.8
simplegeneric==0.8.1
six==1.10.0
sortedcontainers==1.4.4
SQLAlchemy==1.1.11
statsmodels==0.8.0
tables==3.3.0
tensorflow-gpu==1.1.0
testpath==0.3
toolz==0.8.2
tornado==4.5.1
tqdm==4.14.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.12.2
win-unicode-console==0.5
zipline==1.1.1+44.gea677bbb
